pyproject.toml
[tool.poetry]
name = "dimatech-testtask"
version = "0.1.0"
description = ""
authors = ["Dmitry Maslennikov <bake.necko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sanic = "^24.12.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.39"}
alembic = "^1.15.1"
asyncpg = "^0.30.0"
aiomysql = "^0.2.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"
sanic-ext = "^24.12.0"
sanic-testing = "^24.6.0"
pytest-asyncio = "^0.25.3"
factory-boy = "^3.3.3"
deepdiff = "^8.4.2"
click = "^8.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
